move two menu bar settings to setting dialog
